% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_neighbour_dists.R
\name{hex_neighbour_dist}
\alias{hex_neighbour_dist}
\title{[!] Calculate distances between neighbour SOM cells}
\usage{
hex_neighbour_dist(X, method = "euclidean", ...)
}
\arguments{
\item{X}{object of class \code{kohonen} (from package \pkg{kohonen}).}

\item{method}{the distance measure to be used. This must be one of
    \code{"euclidean"}, \code{"maximum"}, \code{"manhattan"},
    \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.
    Any unambiguous substring can be given.}

\item{...}{Further parameters to be passed to \code{\link[stats]{dist}}.}
}
\value{
Data frame with distances between neighbour cells. Data frame has
 these variables: \cr
 \code{Cell_ID1, Cell_ID2} ID codes of neigbour cells. IThese variables are
 sorted so that condition \code{Cell_ID1 > Cell_ID2} is always true. \cr
 \code{Distance} distance between indicated neighbours. \cr
 Other variables are coordinates between neigbourhood boundaries: bedinning,
 end and center points.
}
\description{
[!] Calculate distances between neighbour SOM cells
}
\examples{
library(kohonen)

som_model <- som(Spectra2$spc, grid = somgrid(7,7,"hexagonal"))
hex_neighbour_dist(som_model)  \%>\% head


}
\author{
Vilmantas Gegzna
}
\seealso{
Other SOM.and.hexagon.related.functions.in..pkg.spHelper.: \code{\link{hex_neighbourhood_line_pos}},
  \code{\link{hexagonPT_verticles}},
  \code{\link{hexagonPT}}
}

