% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_neighbourhood_line_pos.R
\name{hex_neighbourhood_line_pos}
\alias{hex_neighbourhood_line_pos}
\title{[!] Neighbourhood line positions of hexagonal cells in map of hexagons}
\usage{
hex_neighbourhood_line_pos(x, y, r = NULL, l = NULL, ...)
}
\arguments{
\item{x}{hexagon center coordinates.}

\item{y}{hexagon center coordinates.}

\item{r}{radius of inscribed circle.}

\item{l}{distance from the center to the upper/lower vertices of hexagon.}

\item{...}{further parameters to methods.}
}
\value{
dataframe with variables \cr
      \code{Cell_ID} - ID number of SOM cell;\cr
      \code{x1, y1, x2, y2} - coordinates of starting and end points of
      neighbourhood boundary lines;\cr
      \code{x_center, y_center} - coordinates of neighbourhood
                                  boundary line centers.\cr
}
\description{
[!] Neighbourhood line positions of hexagonal cells in map of hexagons
}
\examples{
#====================================================================

data <- class::somgrid(8,6,"hexagonal")$pts
head(data)
x <- data[,"x"]
y <- data[,"y"]


hex_neighbourhood_line_pos(x, y)  \%>\% head

#====================================================================

}
\author{
Vilmantas Gegzna
}
\seealso{
Other SOM.and.hexagon.related.functions.in..pkg.spHelper.: \code{\link{hex_neighbour_dist}},
  \code{\link{hexagonPT_verticles}},
  \code{\link{hexagonPT}}
}

