% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon_default_params.R
\name{hexagon_default_params}
\alias{hexagon_default_params}
\title{[!] Calculate parameters for (regular) hexagon}
\usage{
hexagon_default_params(r = NULL, l = NULL, x = NULL, y = NULL)
}
\arguments{
\item{r}{radius of circle inscribed in hexagon. Default value is \code{0.5}.}

\item{l}{distance from the center to the upper/lower vertices of hexagon.}

\item{x, y}{numeric vectors with hexagon center coordinates in hexagon grid.}
}
\value{
List with calculated values of parameters \code{r} and \code{l}.
}
\description{
Function calculates parameters \code{l} and \code{r} for hexagon.
}
\details{
Function assumes regular pointy topped hexagon. (More on geometry of hexagons
 can be found \href{http://www.redblobgames.com/grids/hexagons/}{here}).
Relation between parameters \code{l} and \code{r} in this kind of hexagon: \cr
\code{l = r / cos(pi/6)} \cr
\code{r = l * cos(pi/6)} \cr \cr

These proportions can be distorted to mach grid of x and y values so that
hexahons vere plotted without gaps.
}
\examples{

hexagon_default_params()

hexagon_default_params(r = 1)

hexagon_default_params(l = 1)

data <- class::somgrid(8,6,"hexagonal")$pts
hexagon_default_params(x = data[,"x"], y = data[,"y"])

hexagon_default_params(x = data[,"x"], y = data[,"y"]*2)

hexagon_default_params(x = data[,"x"]*2, y = data[,"y"])

}
\author{
Vilmantas Gegzna
}

