% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonPT_verticles.R
\name{hexagonPT_verticles}
\alias{hexagonPT_verticles}
\title{[!] Calculate coordinates of pointy top (PT) hexagon verticles}
\usage{
hexagonPT_verticles(x, y, r = NULL, l = NULL)
}
\arguments{
\item{x}{hexagon center coordinates.}

\item{y}{hexagon center coordinates.}

\item{r}{radius of circle inscribed in hexagon. Default value is 0.5.}

\item{l}{distance from the center to the upper/lower vertices of hexagon.}
}
\value{
Data frame with coordinates x and y (rows) of 6 verticles as
 variables: A (top) ,           B (left upper),
            C (left lower),     D (bottom),
            E (right) lower and F (right upper).
}
\description{
Calculate coordinates of each pointy top (PT) hexagon verticle
(A, B, C, D, E, F) in respect to center coordinates O(x,y).
}
\examples{

hexagonPT_verticles(1,1)

hexagonPT_verticles(1,1, r = 1)

hexagonPT_verticles(1,1, l = 1)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other SOM.and.hexagon.related.functions.in..pkg.spHelper.: \code{\link{hex_neighbour_dist}},
  \code{\link{hex_neighbourhood_line_pos}}
}

