% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon_default_rl.R
\name{hexagon_default_params}
\alias{hexagon_default_params}
\title{[!] Calculate parameters for (regular) hexagon}
\usage{
hexagon_default_params(r = NULL, l = NULL, x = NULL, y = NULL)
}
\arguments{
\item{l}{}
}
\value{
List with calculated values of parameters \code{r} and \code{l}.
}
\description{
[!] Calculate parameters for (regular) hexagon
}
\examples{

hexagon_default_params()

hexagon_default_params(r = 1)

hexagon_default_params(l = 1)

}

