% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonMap.R, R/hexagonPT.R
\name{hexagonMap}
\alias{hexagonMap}
\alias{hexagonPT}
\title{Plot a pointy toped hexagon and a map of hexagons}
\usage{
hexagonMap(x, y, col = NA, border = "black", r = NULL, l = NULL,
  add = FALSE, ...)

hexagonPT(x, y, col = NA, border = "black", r = 0.5, l = r/cos(pi/6),
  ...)
}
\arguments{
\item{x, y}{hexagon center coordinates.}

\item{col}{the color for filling the polygon. The default,
     \code{NA}, is to leave polygons unfilled, unless \code{density} is
     specified.  (For back-compatibility, \code{NULL} is equivalent to
     \code{NA}.)  If \code{density} is specified with a positive value
     this gives the color of the shading lines.}

\item{border}{the color to draw the border.  The default, \code{NULL},
    means to use \code{\link{par}("fg")}.  Use \code{border = NA} to
    omit borders.

    For compatibility with S, \code{border} can also be logical, in
    which case \code{FALSE} is equivalent to \code{NA} (borders omitted)
    and \code{TRUE} is equivalent to \code{NULL} (use the foreground colour),
  }

\item{r}{radius of circle inscribed in hexagon. Default value is 0.5.}

\item{l}{distance from the center to the upper/lower vertices of hexagon.}

\item{...}{graphical parameters such as \code{xpd}, \code{lend},
    \code{ljoin} and \code{lmitre} can be given as arguments.}
}
\value{
Ploted hexagon
}
\description{
\code{hexagonPT} plots a pointy topped hexagon. \cr
\code{hexagonMap} plots a map of hexagons.
}
\examples{

data <- class::somgrid(8,6,"hexagonal")$pts
head(data)
x <- data[,"x"]
y <- data[,"y"]

hexagonMap(x, y)


hexagonMap(x, y, col = "yellow")


hexagonMap(x, y,
     col = c("tomato1","skyblue","orange"),
     border = c("red", "black"))

text(x,y,1:length(x))
title("Hexagonal cells \\n with cell ID number")


#====================================================================
plot.new()
hexagonPT(.5,.5, r = .25)

#====================================================================

}
\author{
Vilmantas Gegzna
}

